Rename numpy.ipynb to 1_numpy.ipynb
NumPy is the fundamental package for scientific computing with Python.  It is useful linear algebra, Fourier transform, and random number capabilities. Besides its obvious scientific uses, NumPy can also be used as an efficient multi-dimensional container of generic data. Arbitrary data-types can be defined. This allows NumPy to seamlessly and speedily integrate with a wide variety of databases.